# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A production-ready REST API for restaurant menu management",
        "title": "Restaurant Menu API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.restaurant-menu-api.com/support",
            "email": "support@restaurant-menu-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/subcategories": {
            "get": {
                "description": "Get all subcategories with optional filtering and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "List all subcategories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by active status",
                        "name": "active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search in name and description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to order by",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order direction (ASC/DESC)",
                        "name": "order_dir",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include total count",
                        "name": "include_count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.SubCategory"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new menu subcategory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Create a new subcategory",
                "parameters": [
                    {
                        "description": "SubCategory data",
                        "name": "subcategory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_handlers.CreateSubCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.SubCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/subcategories/{id}": {
            "get": {
                "description": "Get a specific subcategory by its ID with optional items",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Get subcategory by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubCategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Include items in response",
                        "name": "include_items",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.SubCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing menu subcategory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Update a subcategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubCategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "SubCategory data",
                        "name": "subcategory",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_handlers.UpdateSubCategoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.SubCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a menu subcategory by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Delete a subcategory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubCategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/subcategories/{id}/display-order": {
            "patch": {
                "description": "Update the display order of a subcategory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Update subcategory display order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubCategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Display order data",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_handlers.UpdateDisplayOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.SubCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/subcategories/{id}/toggle-active": {
            "patch": {
                "description": "Toggle the active status of a subcategory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SubCategories"
                ],
                "summary": "Toggle subcategory active status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "SubCategory ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.SubCategory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_interfaces_handlers.CreateSubCategoryRequest": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "internal_interfaces_handlers.UpdateDisplayOrderRequest": {
            "type": "object",
            "required": [
                "display_order"
            ],
            "properties": {
                "display_order": {
                    "type": "integer"
                }
            }
        },
...
        "internal_interfaces_handlers.UpdateSubCategoryRequest": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "restaurant-menu-api_internal_domain_entities.SubCategory": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "category": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.Category"
                        }
                    ]
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "display_order": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/restaurant-menu-api_internal_domain_entities.Item"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "restaurant-menu-api_pkg_response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "restaurant-menu-api_pkg_response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/restaurant-menu-api_pkg_response.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/restaurant-menu-api_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}